requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
augly==0.2.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorlog==6.6.0
cupy-cuda116==10.3.0
cycler==0.11.0
Cython==0.29.28
docopt==0.6.2
einops==0.4.1
fastrlock==0.8
fonttools==4.29.1
ftfy==6.1.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
Hydra==2.5
hydra-core==1.1.1
idna==3.3
imageio==2.16.1
imgaug==0.4.0
importlib-metadata==4.11.2
importlib-resources==5.6.0
iopath==0.1.9
joblib==1.1.0
kiwisolver==1.3.2
libsvm-official==3.25.0
Markdown==3.3.6
matplotlib==3.5.1
networkx==2.7.1
nlpaug==1.1.3
numpy==1.22.2
oauthlib==3.2.0
omegaconf==2.1.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pipreqs==0.4.11
pkg_resources==0.0.0
portalocker==2.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
python-magic==0.4.25
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scikit-video==1.1.11
scipy==1.8.0
Shapely==1.8.1.post1
six==1.16.0
sklearn==0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2022.2.9
torch==1.10.2+cu113
torch-ema==0.3
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
vidgear==0.2.5
wcwidth==0.2.5
Werkzeug==2.0.3
yarg==0.1.9
zipp==3.7.0
